+++ httk/R/solve_3comp.R
@@ -97,8 +97,14 @@
 #' params <-parameterize_3comp(chem.cas="80-05-7")
 #' solve_3comp(parameters=params)
 #' 
+#'
+#' @seealso \code{\link{solve_model}}
+#'
+#' @seealso \code{\link{parameterize_3comp}}
+#'
+#' @seealso \code{\link{calc_analytic_css_3comp}}
+#'
 #' @export solve_3comp
-#' @useDynLib httk
 solve_3comp <- function(chem.name = NULL,
                     chem.cas = NULL,
                     dtxsid = NULL,