+++ httk/R/solve_model.R
@@ -123,6 +123,7 @@
 #' @useDynLib httk
 #'
 #' @importFrom deSolve ode
+#' @importFrom purrr compact 
 solve_model <- function(chem.name = NULL,
                     chem.cas = NULL,
                     dtxsid = NULL,
@@ -144,13 +145,10 @@ solve_model <- function(chem.name = NULL,
                     method="lsoda",rtol=1e-8,atol=1e-12,
                     recalc.blood2plasma=FALSE,
                     recalc.clearance=FALSE,
+                    restrictive.clearance=TRUE,
                     adjusted.Funbound.plasma=TRUE,
                     minimum.Funbound.plasma=0.0001,
-                    parameterize.arg.list=list(
-                      default.to.human=FALSE,
-                      clint.pvalue.threshold=0.05,
-                      restrictive.clearance = TRUE,
-                      regression=TRUE),
+                    parameterize.arg.list=list(),
                     ...)
 {
 #R CMD CHECK throws notes about "no visible binding for global variable", for
@@ -317,14 +315,14 @@ specification in compartment_units for model ", model)
 # necessarily need all parameters associated with a given model to do this:)
   if (is.null(parameters))
   {
-    parameters <- do.call(parameterize_function,c(list(
+    parameters <- do.call(parameterize_function,args=purrr::compact(c(list(
       chem.cas=chem.cas,
       chem.name=chem.name,
       dtxsid=dtxsid,
       species=species,
       suppress.messages=suppress.messages,
       adjusted.Funbound.plasma=adjusted.Funbound.plasma,
-      minimum.Funbound.plasma=minimum.Funbound.plasma),parameterize.arg.list)) 
+      minimum.Funbound.plasma=minimum.Funbound.plasma),parameterize.arg.list))) 
   } else {
     if (!all(param_names %in% names(parameters)))
     {
@@ -374,7 +372,7 @@ specification in compartment_units for model ", model)
   
   # If the hepatic metabolism is not slowed by plasma protein binding (non-
   # restrictive clearance)  
-  if (!parameterize.arg.list$restrictive.clearance) parameters$Clmetabolismc <- 
+  if (!restrictive.clearance) parameters$Clmetabolismc <- 
     parameters$Clmetabolismc / parameters$Funbound.plasma
   
   # If there is not an explicit liver we need to include a factor for first-
@@ -427,7 +425,8 @@ specification in compartment_units for model ", model)
     non.state.vars <- which(!(names(initial.values)%in%state.vars))
     # Check if there are any non.state.vars
     if(length(non.state.vars)!=0){
-      warning("Additional unnecessary elements were included in the ",
+      if (!suppress.messages)
+        warning("Additional unnecessary elements were included in the ",
               "initial.values -- namely ",
               paste(names(initial.values)[non.state.vars],collapse = ", "),
               ".\n    ",
@@ -447,7 +446,8 @@ specification in compartment_units for model ", model)
       initial.value.units <- compartment_units[names(initial.values)]
       
       #provide warning of assumption
-      warning("No units were specified for the provided initial values. ",
+      if (!suppress.messages)
+        warning("No units were specified for the provided initial values. ",
               "Assume initial values are in default model units:\n    ",
               paste(paste(
                 names(initial.value.units),
@@ -490,7 +490,8 @@ specification in compartment_units for model ", model)
         
         #Check if there are any non.state.vars
         if(length(non.state.vars)!=0){
-          warning("Additional unnecessary elements were included in the",
+          if (!suppress.messages)
+             warning("Additional unnecessary elements were included in the",
                   "initial.values -- namely ",
                   paste(names(initial.value.units)[non.state.vars],collapse = ", "),
                   ".\n    ",
@@ -791,7 +792,6 @@ specification in compartment_units for model ", model)
 # Document the values produced by the simulation:  
   if(!suppress.messages)
   {
- 
     if (is.null(chem.cas) & is.null(chem.name) & is.null(dtxsid))
     {
 # If only a parameter vector is given it's good to warn people that they